Adapted kinetic image model
