--- glotaran/builtin/models/kinetic_image/kinetic_image_megacomplex.pyi
@@ -1,15 +1,0 @@
-from __future__ import annotations
-
-from typing import Any
-
-from glotaran.model import model_attribute
-from glotaran.parameter import Parameter
-
-class KineticImageMegacomplex:
-    @property
-    def k_matrix(self) -> list[str]: ...
-    @property
-    def scale(self) -> Parameter: ...
-    def full_k_matrix(self, model: Any | None = ...): ...
-    @property
-    def involved_compartments(self): ...
