+++ glotaran/builtin/models/kinetic_image/kinetic_image_model.py
@@ -1,10 +1,13 @@
 from glotaran.builtin.models.kinetic_image.initial_concentration import InitialConcentration
 from glotaran.builtin.models.kinetic_image.irf import Irf
 from glotaran.builtin.models.kinetic_image.k_matrix import KMatrix
+from glotaran.builtin.models.kinetic_image.kinetic_baseline_megacomplex import (
+    KineticBaselineMegacomplex,
+)
+from glotaran.builtin.models.kinetic_image.kinetic_decay_megacomplex import KineticDecayMegacomplex
 from glotaran.builtin.models.kinetic_image.kinetic_image_dataset_descriptor import (
     KineticImageDatasetDescriptor,
 )
-from glotaran.builtin.models.kinetic_image.kinetic_image_megacomplex import KineticImageMegacomplex
 from glotaran.builtin.models.kinetic_image.kinetic_image_result import (
     finalize_kinetic_image_result,
 )
@@ -20,8 +23,11 @@ from glotaran.model import model
         "irf": Irf,
     },
     dataset_type=KineticImageDatasetDescriptor,
-    megacomplex_types=KineticImageMegacomplex,
-    #  matrix=kinetic_image_matrix,
+    default_megacomplex_type="kinetic-decay",
+    megacomplex_types={
+        "kinetic-decay": KineticDecayMegacomplex,
+        "kinetic-baseline": KineticBaselineMegacomplex,
+    },
     model_dimension="time",
     global_dimension="pixel",
     grouped=False,
